+++ docs/changes.rst
@@ -40,6 +40,7 @@ Released on ? 2015.
 * Dropped dependency on msginit for creating new Gettext po files.
 * Added configurable dashboard views.
 * Improved notifications on parse erorrs.
+* Added option to import components with duplicate name to import_project.
 
 weblate 2.4
 -----------